third_party/WebKit/LayoutTests/editing/deleting/delete-contiguous-ws-001.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
    assert_selection(
        '<div contenteditable><span>foo    ^bar|   \nbaz</span></div>',
        'delete',
        '<div contenteditable><span>foo\u{00A0}| baz</span></div>',
        'foo');
}, 'Delete collapses whitespaces into U+00A0');
</script>
